Reorganised KeyResource #2765
