+++ src/lib/platform/CMakeLists.txt
@@ -18,8 +18,8 @@ if (WIN32)
 	file(GLOB headers "MSWindows*.h")
 	file(GLOB sources "MSWindows*.cpp")
 elseif (APPLE)
-	file(GLOB headers "OSX*.h" "*KeyResource.h")
-	file(GLOB sources "OSX*.cpp" "OSX*.m" "*KeyResource.cpp")
+	file(GLOB headers "OSX*.h")
+	file(GLOB sources "OSX*.cpp" "OSX*.m")
 elseif (UNIX)
 	file(GLOB headers "XWindows*.h")
 	file(GLOB sources "XWindows*.cpp")
