+++ src/lib/synergy/IKeyResource.h
